removed default react files, moved Parkbook.js
