+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter as Router } from "react-router-dom"
-import { Parkbook } from './Parkbook'
+import { Parkbook } from './components/Parkbook'
 
 ReactDOM.render(
   <React.StrictMode>
